Testing EVM processing.
In this changeset I am testing how the EVM would work in our project. So far everything works perfectly.
The bulk of the testing happens in vm::tests::smoke(), where two things happen:

1. First a contract creation bytecode straight from solidity compiler is executed against the EVM which results in an output that is the final deployed contract code.

2. Then I am using the deployed contract code and executing a transaction against it that calls store(12). The desired output is that the address location of the contract is accessed to determine if it is cold or warm storage (EIP 2939), then a value of 12 is stored there at slot 0x0 (the first and only variable in the contract).

3. Running `cargo test -- --nocapture` will output the correct and expected console outputs for every executed instruction and state access.

Next: Turn this into a real unit test, that will serve as a smoke test for evmodin, to make sure that it is working as expected between version upgrades.